timber: tree: simplify node removal
Our client tree maintains the invariant that for every node, it either
has two children (in which case it is splitting available space) or it
has a client. Due to that, when removing a node from the tree, then its
parent can only either be the tree's root node or a parent with two
children. In the second case, we only need to adjust this immediate
parent by making it a leaf-node with a client only.