timber: desktop: correctly update linked list when inserting new desktop
When inserting a new desktop, then we insert it between the currently
focussed desktop and its next desktop. This requires us to update the
`next` pointer of the current desktop as well as updating the `prev`
pointer of the current next desktop. While we do update the former one,
we fail to update the next desktop correctly, which leads to faulty
ordering.

Fix this issue by correctly updating both pointers.