timber: events: remove return codes from event handlers
The retrun codes aren't used at all right now, and in fact the event
handlers are left to decide themselves whether a potential error is
fatal or not. Thus, we can remove the return codes altogether to
simplify the code and to not pretend that they were meaningful in any
way.