timber: desktop: fix input focus when destroying client
When destroying clients on a desktop whose screen is currently not
focussed, then we try to adjust the input focus anyway, ending up with
the input focus residing on that other screen. Fix this by only setting
the input focus in case where a client is getting destroyed on the
currently focussed screen and desktop.