timber: tree: set new node's split according to their ratio
When inserting a new client into the tree, we currently always split the
client's parent tree vertically. While simple and deterministic, we can
try to make better use of the available space. Instead of always
splitting the same, we can split according to the aspect ratio of the
node's available space:

- if the node is wider than it is high, then we split vertically
- if the node is higher than it is wide, then we split horizontally