timber: client: stack non-fullscreen windows below others
When setting a client to fullscreen mode, then we set its window to
stack above others so that it is not obstructed. When resetting that
client to tiled mode again, we will not revert its stacking mode but
again set it to stack above others.

Fix this by stacking the window below others if setting a client to
non-fullscreen mode.