timber: fix input focus when desktop is empty
When there's multiple desktops and one of both desktops is empty, then
the input focus will be set incorrectly when selecting the empty
desktop. Instead of not having any input focus at all, the input focus
will be set to the window that's located below the current pointer
position.

X11 does not allow for having no input focus at all for non-reparenting
window managers. To work around this, we create a dummy window that we
will set the input focus to instead of setting it to the root window. To
avoid any window manager messing with this dummy window, we set the
override redirect window attribute on it.

This fixes the described input focus issues.