timber: events: do not invoke event handler recursively
When re-layouting clients on a desktop, we need to ignore all
"XCB_ENTER_NOTIFY" events as otherwise the current focus would
inadvertently change to where the mouse pointer currently is. To do so,
we fetch all outstanding events and then manually run the event loop
once until no more events are pending, ignoring all "XCB_ENTER_NOTIFY"
events. This is done recursively inside `tmbr_desktop_layout`, which has
several downsides:

- There's potentially unbounded recursion going on, which may overflow
  available stack space.

- The program flow is kind of awkward to understand when nested events
  occur.

- We trigger and handle other events while currently still processing
  the current event, which may lead to an inconsistent intermediate
  state.

The most important point is the last one, and it already causes a `NULL`
pointer dereference right now when new clients are being added that
immediately send a fullscreen message:

    1. A new client is being added to the current desktop. We add it to
       the desktop's tree and then re-layout it, which causes the nested
       event handler to run.

    2. The client message requesting fullscreen access is being
       processed. At this point, we haven't yet adjusted the desktop's
       focus, as the previous event handler hasn't finished yet.

    3. We set the fullscreen flag of the desktop and try to re-layout it
       again. Due to the fullscreen flag being set, we will directly try
       to move the current focus such that it covers the complete
       desktop.

The third step here is the important one. In step 1, we haven't yet set
the focus, but we now try to set the focussed client to fullscreen mode.
If the desktop has other clients already, the wrong client will be set
to fullscreen mode. Otherwise, focus will be a `NULL` pointer and we
run into a segfault.

Fix this avoiding recursive event handling. Instead of manually running
the event loop in `tmbr_desktop_layout`, we just set a new global state
variable `ignored_flags`. This flag, if set, will cause us to ignore all
events matching it in the current X11 event loop. After we have
processed all events, the flag will be reset such that we again start to
honor all events.