wm: rename `fifofd` to `ctrlfd`
The control FIFO file descriptor is currently stored in the
`fifofd` field of our global state. As we're about to refactor
our code to instead use a Unix socket, this field will be
misnamed. Thus, let's rename it up front to `ctrlfd`, which is a
better name even now without the refactoring.