ctrl: use AF_UNIX sockets
Since timber's inception, we've been using a FIFO control pipe to
receive commands. The downside of using FIFOs is that they are
unidirectional only and thus do not allow us to return any data
to the client. There are some usecases though which would require
us to do exactly that, for example if we want to be able to
return events to registered clients or success/failure messages.

Switch over client and window manager to use AF_UNIX sockets in
SOCK_STREAM mode. We do not yet make any use of the new
possibilities, though, but will do that at a later point.