wm: ignore enter events on unmaps
When unmapping a window, the X server will generate an enter
event if the mouse curser has been located inside of that window
and another window is located beneath. This will cause timber to
adjust focus, but in fact we do not want to adjust focus but
simply keep the old window focussed and afterwards optionally
re-focus if the old focus is about to be destroyed.

Ignore all enter events when receiving an unmap event to fix the
described issue.