common: implement control socket read/write functions
When writing to the control socket of the window manager, we
neither verify that we were able to retrieve the complete command
on the window manager side nor are we verifying that the command
has been parsed and executed correctly on the client side. To
implement this, we need to be able to exchange multiple messages
between WM and client, but this in turn requires us to be able to
determine message boundaries.

Implement new functions `tmbr_ctrl_read` and `tmbr_ctrl_write`.
For now, those functions are implemented in the easiest way
possible: we will always write commands with a terminating NUL
byte and read commands characterwise until we find the first NUL
byte.