wm: have commands return proper error codes
In order to be able to make errors available to the client, we need the
window manager to be able to see whether a command was successful or
not. Thus we change the return type of all command handlers to return an
error code that indicates whether the command was successfully executed
or not. For now, we use standard POSIX error codes, even though these
might not always perfectly fit the bill.