cmd: implement client return codes
In order to make the client scriptable, one needs to determine whether a
command was successfully executed or not. As we have the error
information available in the window manager now, we can modify the
protocol between window manager and client to always expect back an
error code after executing a command. Depending on whether the code
indicates success or not, we can alter the return value of the client
and generate error messages.