wm: do not die when commands cannot be read
When encountering any unexpected errors while reading from the control
socket, we have been historically dying instead of simply ignoring the
command. As the previous control implementation was using a FIFO, it was
impossible to recovery from reading errors in some scenarios, which is
why it made sense back then. But now with our Unix sockets we may simply
close the client socket and continue chugging along.