wm: fix toggling fullscreen for unfocussed clients
When a client requests to go into fullscreen mode, then we adjust the
`fs` bit of its desktop and re-layout it. If the message is sent by an
unfocussed client, though, then we will notify it of being set to
fullscreen, but will draw the currently focussed client at full size
instead of the correct one.

Fix this by adjusting the desktop's focussed client before re-layouting
windows.