ctrl: fix warning on uninitialized read
Coverity complains about `tmbr_ctrl_read` potentially reading an
uninitialized element of the buffer that is being passed in. This is in
fact impossible to trigger, as we know that read(3P) has just written
to the exact position we're about to read. But to make Coverity happy,
we can re-write the loop to read everything into a local `char` variable
and overwrite the buffer manually.