wm: fix killing of last client on desktop
When killing the last remaining client on the current desktop, then
there will be no sibling remaining to which we can adjust the focus to.
Thus, a failing call to `tmbr_tree_find_sibling` is not an error, but
perfectly fine as we should just set the desktop's focus to `NULL` in
that case. Still, since commit 146a952 (wm: check return value of
`tmbr_desktop_focus`, 2019-07-21) we're raising an error if the function
returned an error.

Fix this by treating an error of `tmbr_tree_find_sibling` as valid.