wm: desktop: fix swap command
The swap command is implemented by simply swapping the client trees
between focussed and selected desktop. This forgets though that clients
themselves have a pointer to the containing desktop, which wasn't
adjusted properly and thus continued to point to the wrong desktop.

Fix this by implementing a proper list swap. Instead of swapping
clients, we now change positions of the desktops in the doubly linked
list.