wm: manage clients' WM_STATE property
The WM_STATE property defined by ICCCM is used by clients to get to know
whether they are treated by the window manager to be either in the
"normal", "iconic" or "withdrawn" state. The information can then be
used to know e.g. whether the client should animate its window, its
icon or whether it is about to be withdrawn from the window manager.

The field is currently not implemented at all. While not noticeable for
most clients, some may behave erroneously. One example is WINE: when
a WINE window is about to unmap, it will send a UNMAP_NOTIFY message to
the WM and then wait until the WM has put it into "withdrawn" state. If
that state is not reached, then it will timeout after some few seconds.

Implement the WM_STATE property. For timber, only the states "normal"
and "withdrawn" are of importance, as there are no iconified windows.
Thus, we now set adopted windows to "normal" mode and will set them to
"withdrawn" when receiving UNMAP_NOTIFY events.