wm: client: support WM_TAKE_FOCUS protocol
Setting the input focus via xcb_set_input_focus(3) does not notify the
target window that it is now the currently focussed one. Thus, ICCCM has
introduced the WM_TAKE_FOCUS protocol. If a window publishes the
WM_TAKE_FOCUS atom as part of WM_PROTOCOLS, then the window manager
shall send it a client message of type WM_PROTOCOLS with WM_TAKE_FOCUS
as value to instruct it to take focus. This allows clients to perform
additional actions upon becoming focussed.

This commit implements the described protocol to improve interaction
with clients that expect it.