wm: events: fix stale event mask causing missed events
When an event mask is currently active, then all events queued up that
match the mask will be ignored. To not indefinitely ignore such events,
the mask is being reset when no events are currently queued anymore.
This fails to account for event masks set up by commands, though. As
commands are processed after events, the event mask will not get reset.
Thus, the event mask will remain until the next batch of commands has
been processed.

Fix the issue by reordering command and event handling. Now, we will
always handle commands first, so that any events that shall be ignored
as a result of a command will be queued up. Afterwards, we will handle
these events and potentially ignore any that match the event mask.
Furthermore, we now always reset the event mask independent of whether
any events were processed or not.