wm: events: properly handle events generated by command
If a command wants to ignore a certain type of events generated by its
actions, then it sets up the event mask and calls xcb_aux_sync(3) to
pull in the events. In case where there has been no pending events
previous to the call to xcb_aux_sync(3), though, the main loop will not
handle events at all as the call to poll(3P) will not have set the
POLLIN flag for the XCB file descriptor. This may result in the event
mask not taking effect at all.

Fix the issue by unconditionally calling `xcb_poll_for_event` in the
main loop.