wm: fix focus change when unmapping unmanaged windows
When windows get unmapped, they may uncover underlying windows that they
have previously been obstructing. In case the unmapped window is an
unmanaged window, we thus need to ignore EnterNotify events in order to
not act like the mouse just moved into the now-uncovered window and to
retain the current focus. While the code already wants to do this, we do
in fact exit early if we notice that the unmapped window is not managed
by us. As a result, we fail to correctly set up the ignored events to
start ignoring EnterNotify.

Fix the issue by always executing the event synchronization and setup of
the ignored event.