ctrl: unify connecting to control socket
The current logic to assemble the control socket's path and to connect
to it is split up between client and window manager. As both locations
share some common code to set up the socket itself and to compute the
path of the socket, it makes a lot of sense to unify them both into a
single location.

Implement a new function `tmbr_ctrl_connect` that unifies the code.