common: extract `read_bytes` function
Extract a function `read_bytes` that works similarly to `write_bytes`.
It will read an exact amount of bytes and return an error if there's
less bytes available, handling both EAGAIN and EINTR. The sole user is
currently `tmbr_ctrl_read`, for which it doesn't really gain a lot as
it's always reading byte-wise anyway. As such, this is a preparatory
commit only that will aid in implementing PKTLINE-style packets.