common: introduce control packet types
In order to be able to discern different types of packages, we need to
be able to send information about the respective packet types across the
wire. To do so, we now introduce a prefix-encoding, where each prefix
consists of

    - 1 byte to encode the packet type
    - 4 bytes to encode the message length

This allows us to send more information than currently, where we only
send an error code and nothing else.