tmbr: client: discern wrong packet types and server errors
Right now we print the same error message in case we've got a wrong
packet type or a server error as a command's response. In case of an
unexpected packet type, though, we can't generally assume that its
message will parse as an error code, making the resulting error message
wrong and confusing.

Fix the issue by splitting up both failure cases and use different error
messages for each of them.