tmbr: wm: ignore SIGPIPE signals
When writing to a client socket, it may cause that we get a SIGPIPE due
to the socket having become invalid in case a client disconnects or
otherwise fails. Default behaviour on Unix-like signals is to raise a
SIGPIPE signal in that case, which currently causes timber to fail as we
do not handle the signal at all.

Fix the issue by ignoring SIGPIPE. As we're handling I/O errors anyway,
we do not want to have the OS cause us to abort. Furthermore, we need to
handle SIGPIPE for the upcoming event notification mechanism, where we
keep track of potentially long-lived file descriptors.