tmbr: ctrl: always expect data before final error package
Right now we first expect data packages after an initial error package
which signals no error. This was done for the subscription service,
which was expected to first get a positive acknowledgement and then
afterwards process data. It turns out this is rather inflexible, though,
when implementing other commands that consume data.

So let's instead switch order to expect data first and then finally
expect an error package to notify us we're done.