wm: include command errors in reported events
When executing a command, timber currently will notify any subscribers
to the fact that the command is about to execute, but not what its
result is. Most importantly, if an error occurred turing command
execution, we won't ever tell subscribers.

Fix this by including the error code in the event.