timber: return error code from `tmbr_wm`
When no arguments are passed to timber, we assume the window manager is
to be run and thus execute `tmbr_wm()` directly, but without returning
its return value. As a result, the client-side code will run immediately
afterwards when the window manager exits.

Fix the issue by returing `tmbr_wm`'s return code.