common: send packet structure instead of serializing data
When sending commants to the server, we currently parse the data twice:
first on the client-side and then again on the server-side. This is
inefficient and makes it harder than necessary to extend commands with
additional argument, as one always has to implement parsing,
serialization and deserialization.

Let's improve the situation by always sending serialized data over the
wire. To do so, this commit extends the `tmbr_pkt_t` structure to embed
all possible arguments which can be sent to the server in a tagged
union. Afterwards, we'll directly send the complete union instead of a
size-prefixed string, which greatly simplifies both the sending as well
as the receiving side.

As a result, we can now send over typed data and can ged rid of
server-side parsing of received data and command arguments.