wm: ctrl: fix recursive timber instances clobbering control
As we're exporting `TMBR_CTRL_PATH` after we've set up the control path,
we'll also override the control socket in case we're spawning recursive
timber instances.

Fix the issue by not exporting `TMBR_CTRL_PATH`. We're already exporting
`WAYLAND_DISPLAY` which, in combination with `XDG_RUNTIME_DIR`, is
enough to identify the new socket path.