common: ctrl: implement packet tags
It's currently easily possible to send packets over the wire which
are impossible to interpret by the other side, e.g. if using an
incompatible client. This may go unnoticed by the server, which would
then start to interpet the packet in the wrong way.

Make the issue more unlikely to happen by always inlucding a trailing
tag. If the tag doesn't match, the packet will ignore.