wm: binding: fix bindings being level dependent
In order to find the correct binding for a given keycode and its
modifiers, we need to translate it from the current keycode, which is
level-dependent, to its level-independent version. This is currently
done by always translating it to its lower-case variant, but that only
accounts for the shift level. And even in that case it's quite naive.

Fix the issue by using `xkb_keymap_key_get_syms_by_level`, which allows
us to pick the level we want to apply.