wm: server: implement DPMS via idle timeout
In order to save energy and protect screens, they tend to be turned off
automatically after a certain amount of idle time. As we've now got both
idle and idle inhibitor managers set up, we can make use of their events
in order to automatically turn on and off screens based on user
activity.