wm: screen: don't let hidden clients damage screen
When a client commits, then we damage the screen which is associated
with its desktop even if that desktop isn't currently shown. That
doesn't make a lot of sense considering the damage won't be visible
anyway.

Refine the damage strategy to only damage the screen in case the client
is currently visible.