wm: screen: fix leaking pixman_region32_t when screen's not damaged
When we decide to not render anything as no frame is required, then we
exit early. As it's possible that the `pixman_region32_t` got used by
`wlr_output_damage_attach_render` and thus contains allocated data, we
should make sure to always free it.

Fix the potential leak by adding a common exit path which finalizes the
structure.