wm: desktop: fix switches resetting fullscreen mode
When switching between desktops or screens, we're transitively calling
`tmbr_desktop_focus_client`. One of the effects of this function is that
the desktop's fullscreen mode is reset, which is unexpected behaviour.

Fix the issue by only resetting fullscreen mode if the desktop's current
focus is different to the new focus.