wm: desktop: fix NPE when calculating fullscreen layout
When recalculating the layout for a fullscreen desktop which doesn't
have any clients, then `tmbr_client_set_box` is called with a `NULL`
pointer. As this function doesn't check whether it's being passed a
valid client, this may in turn cause a `NULL` pointer exception.

Fix the issue by only calling `tmbr_client_set_box` if the desktop has a
focus.