wm: client: only render damaged clients
The current damage tracking implementation is quite bare-bones: as soon
as any client has damaged the screen, all clients on the same screen
will get redrawn. It's obvious that this is inefficient and wasteful.

Improve the situation by not damaging the whole screen in most cases,
but only the client's rectangle. When rendering clients we can then
check whether the client's geometry is contained in any of the damaged
regions -- if not, we simply skip redrawing it.