wm: screen: don't damage whole screen when focussing desktop
When calling `tmbr_screen_focus_desktop`, the complete output is
currently getting damaged. While this is required in case we're
switching focus between different desktops, we only need to re-draw the
focussed client in case the desktop is already the current focus.

Improve code to only redraw the complete output in case the desktop
focus is getting switched.