wm: keyboard: tear down keyboards on destroy event
When keyboards get unplugged, the compositor is getting a `destroyed`
event from wlroots. Timber doesn't currently listen to this event, which
will cause it to continue listening on any events on this destroyed
resource and to leak the keyboard's memory.

Fix this by correctly setting up a destroy event listener which
unregisters any listeners and free's the structure.