wm: cursor: fix searching for screen by output
The `wl_list_for_each` macro will never set the iterator to `NULL`, but
instead iterate over the list until the iterator has arrived back at the
list's `HEAD`. It's thus invalid to check whether a certain item was
found in the list by checking for a `NULL` pointer, which is exactly
what's being done when trying to search the screen corresponding to a
given output.

Fix the issue by comparing with the list head instead.