wm: cursor: don't enter cursor at (0,0) when focussing a client
When focussing a client, it always receives a cursor-enter notification
with coordinates (0,0). This will cause the client to think the cursor
is located in its top-left corner, even though it could've been located
somewhere else entirely.

Fix the issue by simply removing the notification. The enter
notification is handled by our cursor motion code anyway, so there's not
much of a point to also notify when focussing the clients.