wm: cursor: don't needlessly damage client on motion
When the cursor moves above a client, then we'll make sure to always
have it focussed, damage it and notify it about the movement. The
second step about explicitly damaging the client is not required though
as it's already getting damaged by focussing it.

Remove the explicit damage call to avoid doing dupicate work.