wm: client: use scissor boxes to not render atop other clients
When a client is resizing or simply not adhering to a resize request
sent by timber, then its geometry may exceed what the layout thinks it
should be. The window manager isn't coping well with this at the moment
and as a result clients may either be rendered too small or atop of
other clients. Especially with per-client damage tracking, this can
cause user-visible artifacts.

This commit improves the situation when rendering a client which exceeds
its boundaries by using a scissor box. Instead of rendering atop of
other clients, we simply forbid any rendering outside of its boundaries.
While the client itself will be cut off, at least other clients won't
get damaged because of it.