wm: screen: load scaled Xcursor theme when output scale changes
We currently only load the Xcursor theme once at setup time with a
default scale of `1`. That's perfectly fine as long as the output scale
doesn't change, but may be suboptimal in case it does.

Improve code by re-loading Xcursor themes with custom scales in case the
output scale change.