wm: cmd: use double forking to spawn commands
When spawning commands, we currently do a fork and let the child process
execute the command directly. This causes some issues when we're the
current session leader, as we are not reaping zombie processes of
grandchildren.

Fix this issue by using a double-forking pattern and making the
immediate child a session leader. Like this, timber stops being
responsible for this process group and instead hands over the
responsibility of reaping processes to PID 1.