wm: setup: use access(3P) to check for config execuatable bit
We're currently using stat(3P) to determine whether a config file exists
so we can skip over it in case it doesn't. We're not checking the
executable bit though, so it could be that it exists but is not
executable.

Improve this by using access(3P) instead with `X_OK`.