ctrl: use custom Wayland protocol to control timber
In order to control timber, we currently implement a custom protocol to
exchange information between client and window manager via serialization
and deserialization of a structure. But in fact, Wayland already
provides all the mechanisms for such custom protocols via global
interfaces, so we've essentially duplicated the functionality.

Rewrite command handling to use a custom Wayland protocol in order to
get rid of the self-built implementation. This also removes the ability
to register for event updates. Clients should instead register for
Wayland events directly.