client: simplify passing of string arguments
Previous to the conversion to a custom Wayland protocol, we were always
serializing and deserializing a structure in order to pass data over the
wire to the compositor. This required us to have explicit length limits
for any strings which are being passed as the strings were copied into
the to-be-serialized struct. This isn't necessary anymore with the
custom Wayland protocol as serialization is handled by the library.

Simplify the code to avoid copying strings. Instead, we now just set up
pointers as required.