proto: remove nearest selection from
The "nearest" selection is supposed to select the neighboring client in
the current node, but it's only supported by `tmbr_tree_find_sibling`.
Exposing it via the protocol is thus weird, as we don't really support
it.

Remove the enumeration to avoid any confusion and inline the code into
the only callsite.