wm: desktop: damage output when changing fullscreen mode
With commit a5f2c35 (wm: desktop: don't damage screen if setting
fullscreen, 2020-10-07), output damaging was removed when changing
fullscreen mode. The change causes clients to not be updated though in
case we're leaving fullscreen mode as the desktop recalculations will
only damage the single fullscreen client, whereas the other clients will
not get damaged.

Fix the issue by again damaging the whole output when changing
fullscreen mode. In contrast to before, this is now only done when
the new fullscreen mode is actually different from the current one.
Before, the output was always damaged regardless of whether a change is
done or not.