wm: desktop: stop resetting fullscreen mode
Commit a3345db (wm: desktop: damage output when changing fullscreen
mode, 2020-10-08) tried again to fix issues with output damaging when
setting or unsetting fullscreen mode. Turns out it's been faulty once
more as callsites rely on the fact that `tmbr_desktop_set_fullscreen`
always recalculates the layout, but the function now returns early in
case no change is done. Which is arguably the right thing as other
callsites shouldn't rely on implicit behaviour.

This commit fixes the issue by changing behaviour altogether: instead of
resetting fullscreen mode on most of the commands, we now keep it as-is.
So if a new client spawns, another one gets focussed et cetera the
current fullscreen state will not be touched. Instead, we now explicitly
invalidate the layout via a call to `tmbr_desktop_recalculate`.