wm: desktop: recalculate layout when focussing clients
When focussing another client, it's not directly obvious that it's
necessary to recalculate the layout, and in fact it isn't in case the
desktop is currently not in fullscreen mode. But when it is, then we
need to resize the newly focussed client to have the screen's dimensions
as we don't do so for all clients of the fullscreen desktop.

Fix this by always calling `tmbr_desktop_recalculate` when focussing a
client.