wm: desktop: fix crash if focussing NULL client
When focussing a desktop without any clients left, we call
`tmbr_desktop_focus_client` with a `NULL` client. This is thus a
perfectly fine scenario, but we still unconditionally dereference the
client even if it may be `NULL`. As a result, this may cause a NULL
pointer reference.

Fix the error by using `desktop` instead of `client->desktop`.