wm: client: clear surface if geom is too small
In case a rendered client doesn't fit into its rectangle, e.g. because
it is maintaining its aspect ratio or because it is currently resizing,
then we simply draw nothing into the empty space. Due to double
buffering, this may cause us to render content of the alternate buffer
into that empty space and cause artifacts.

Fix the issue by clearing the client's rectangle in case its geometry
is too small for its designated rectangle.