wm: screen: avoid calling desktop functions with invalid setup
When focussing a desktop, we first assign the desktop's focus only to
set up the screen's state afterwards. As a result, the desktop may see a
screen with invalid state. One of our core assumptions is that a screen
always has a desktop assigned and that the server always has a screen,
but the current order of assignments invalidates that assumption.

Fix the code by first setting up the screen and server and only
adjust the desktop focus afterwards.