wm: desktop: unfocus current focus when focussing a client
When focussing a new client on a given desktop, then we'll unset focus
on the desktop's current focus. This may work alright in some cases, but
in general we should be unfocussing the global focussed client instead
of the local one. Otherwise it may cause invalid state in case one was
switching focus from a different desktop or screen.

Fix code by always unsetting focus of the global focus.