wm: cursor: clear pointer focus if it's not inside client
The current pointer handling is wrong in case where the pointer is not
inside of the client's rectangle. Instead of notifying the client that
it's outside of its rectangle, we instead just pass some bogus
coordinates to it. This may thus cause us to generate click events at
spurious locations.

Fix the bug by clearing pointer focus in case it's not currently
pointing at any of the client's subsurfaces.