wm: client: extract function to notify about pointer movements
There's currently two places where we handle pointer notifications: once
when a client is being focussed and once when the pointer is being
moved. The logic is non-trivial, and there's going to be a third
callsite in the next commit.

Refactor code and extract a new function `tmbr_client_notify_pointer`
to unify the logic.