wm: client: handle pointer on commit
Currently, we handle pointer notifications on recalculating the desktop
layout. The intention is to notify clients when a resize of their window
caused the pointer to enter it. This doesn't work though as the resize
happens async, and at the point where we send the pointer notification
it likely hasn't finished yet.

Fix the issue by handling pointer notifications on the client's commit
callback.