wm: client: correctly scale damages
When recalculating the desktop layout, we store client extents as
effective resolution. Extents thus don't match the actual output
dimensions, but need to be scaled. While we do correctly scale extents
when rendering, we don't when damaging clients.

This has worked fine until now given that we always redraw the whole
client and perform checks whether a client was damaged in an unscaled
fashion, too. But we're about to implement per-surface damage tracking
and rendering, so let's fix this now by correctly scaling the damages.