wm: client: move damage detection into surface code
The detection whether we need to redraw clients is currently only able
to detect whether a client needs to be redrawn in full. This means that
if we detect a redraw is required, we have to redraw all surfaces of the
client, which may be inefficient when only subsets of them change.

This commit instead moves detection of damages into the surface code.