wm: client: move clearing of damaged regions into screen
We need to clear any damaged regions to not retain any content which
those regions contained before in case they're not getting rendered on
by the client. This is currently done as part of the client's rendering,
but in fact it's independent of any clients and can be done by the
screen.

This commit thus moves the code. This should also be more efficient
considering we now only need to compute those damaged regions in a
single loop. It also allows us to move detection of damages into the
surface rendering code more easily.