wm: client: fix uninitialized access when rendering surface without texture
When a surface either has no texture or when it wasn't damaged, then we
simply skip rendering it. But in the former case of it not having a
surface, we directly exit the function and call `pixman_region32_fini()`
on the not-yet-initialized damage region. We thus potentially free
uninitialized data and thus cause the compositor to crash.

Fix this by only checking for a missing texture after the region was
initialized.