wm: client: take client border into account when damaging
When damaging client surfaces, we need to translate surface-relative
coordinates of the damaged rectangles into absolute coordinates on the
output. We already take into account the client's coordinates and
display scale, but we forgot to take the client's border into account.

Fix this issue by correctly adding the border width to both X any Y
coordinates.