wm: tree: simplify use of tree iteration
The tree iteration requires the caller to pass in three variables: the
tree that is to be iterated, the iterator used to store a temporary
value and the variable which is going to be filled with the tree entries
while we iterate. This is kind of verbose, but also the initialization
of both iterator and iteration value breaks the macro in case it's e.g.
executed in a braceless conditional statement.

Let's fix both issues by converting to a for-loop. Both iterator and
value are initialized by the macro now so that they don't have to be set
by the caller anymore.