wm: desktop: recalculate desktop when changing screen focus
The desktop layout gets calculated on demand whenever a new client gets
added or removed or when its effective resolution changes. But when the
effective resolution changes, only the layout of the current desktop
gets recomputed, and we never recompute it if changing a screen's
focussed desktop. As a result, if the effective resolution changes, then
currently unfocussed desktops do not get recomputed even when getting
focussed.

Fix this issue by recomputing all desktop layouts when the effective
resolution changes.