wm: render: make rendering logic indepentent of client
While the rendering logic can be made to work generically for all
surfaces, it's currently tied to `tmbr_client`s. We're about to
introduce support for the layer-shell protocol though which will require
rendering logic, too.

Refactor the code to be indepentent of the client by passing in the
client's bounds and the respective output instead of the client itself.