wm: render: move frame-done notifications into renderer
While the rendering logic hosts a generic implementation of a renderer,
it doesn't yet handle sending of frame-done events. This makes it harder
than necessary to reuse the infrastructure, as we now have to implement
sending  of these notifications for every time of surface which gets
rendered.

Fix this issue by moving sending of the frame-done notification into the
renderer.