wm: screen: clear scissors when rendering empty screen
When no clients exist on a given desktop, then we need to render an
empty screen. To do so, we just call `wlr_renderer_clear()` directly.
But because it doesn't clear scissors first, the code relies on other
renderers to always clean up scissors. While that's the case for now, it
is fragile and will be changed in a subsequent commit.

Fix the issue by always clearing scissors before clearing the screen.