wm: render: restrict rendering to bounding box
While the rendering logic scissors by using damaged rects, we do not
restrict the scissored parts by the bounding box. This allows a client
to render over its bounding box if any of its surfaces is wide enough to
bust the bounds.

Fix this by intersecting the surface's extents with the client's
bounding box.