timber: require "run" argument to start WM
The window manager is currently started by passing no arguments to the
timber executable. This is awkward given that it's easy to accidentally
start a window manager if one for example simply wants to see what
arguments timber supports.

Fix this by requiring a "run" parameter to start the window manager.