wm: screen: clear damaged rectangles with background color
When rendering damaged clients, we first clear damaged rectangles with
black and then render over them immediately afterwards. This code is
required in order to not leave previously rendered data intact in case
the damaged client somehow ends up not rendering into the damaged rect.

With the upcoming support of the layer shell, this code becomes more
important: it will also be used to clear the exclusive area in case a
layer-shell client doesn't render into it. But In that case, it does
make more sense to render with our background color.

Let's do so and start using the same color.